requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
sqlmodel==0.0.24
asyncpg==0.30.0
cloudinary==1.42.2
python-dotenv==1.0.1
passlib==1.7.4
PyJWT==2.10.1
python-multipart==0.0.20
python-dotenv==1.0.1
bcrypt==4.3.0